// src/app/shared/models/profile.model.ts
export class Profile {
    dob: string;
    userID: string;
    bloodType: string;
    firstName: string;
    lastName: string;
    zipCode: string;
    gender: number;
    pregnant: number;
    birthdate: number;
    height: number;
    weight: number;
    userName: string;
    privateKey: string;
    extraData: any;
    email: string;
    longitude: number;
    latitude: number;
    city: string;
    cellNumber: string;
    licenseNumber?: string;
    npi: string;
    profileEmail: string;
    languageName: string;
    oemID: number;
    imageURL: string;
    screenName: string;
    locationTime: number;
    appVersion: string;
    directEmailAddress: string;
    loopSocialId: number;
    languageId: number;
    sendPushNotification: boolean;
    userKey: string;
    secretCode: string;
    premium: number;
    providerStatus: string;
    subscriptionExpiration: number;
    employerID: string;
    emrID: string;
    passport: string;
    ethnicity: string;
    activated: boolean;
    providerId: string;
    stringDob: string;
    address: string;
    state: string;
    publicKey: string;
    country: string;
    updateDirectEmail?: boolean;
  
    // data(key) {
    //   const _data = JSON.parse(this.extraData)
    //   return _data[key]
    // }
  
    // setData(key, value) {
    //   const _data = JSON.parse(this.extraData)
    //   _data[key] = value
    //   this.extraData = JSON.stringify(_data)
    // }
  }
  